move more tests to test_slice.py
