+++ pandas/tests/indexing/multiindex/test_multiindex.py
@@ -407,50 +407,6 @@ class TestMultiIndexBasic(object):
             with catch_warnings(record=True):
                 df.ix[name, 'new_col'] = new_vals
 
-    def test_multiindex_label_slicing_with_negative_step(self):
-        s = Series(np.arange(20),
-                   MultiIndex.from_product([list('abcde'), np.arange(4)]))
-        SLC = pd.IndexSlice
-
-        def assert_slices_equivalent(l_slc, i_slc):
-            tm.assert_series_equal(s.loc[l_slc], s.iloc[i_slc])
-            tm.assert_series_equal(s[l_slc], s.iloc[i_slc])
-            with catch_warnings(record=True):
-                tm.assert_series_equal(s.ix[l_slc], s.iloc[i_slc])
-
-        assert_slices_equivalent(SLC[::-1], SLC[::-1])
-
-        assert_slices_equivalent(SLC['d'::-1], SLC[15::-1])
-        assert_slices_equivalent(SLC[('d', )::-1], SLC[15::-1])
-
-        assert_slices_equivalent(SLC[:'d':-1], SLC[:11:-1])
-        assert_slices_equivalent(SLC[:('d', ):-1], SLC[:11:-1])
-
-        assert_slices_equivalent(SLC['d':'b':-1], SLC[15:3:-1])
-        assert_slices_equivalent(SLC[('d', ):'b':-1], SLC[15:3:-1])
-        assert_slices_equivalent(SLC['d':('b', ):-1], SLC[15:3:-1])
-        assert_slices_equivalent(SLC[('d', ):('b', ):-1], SLC[15:3:-1])
-        assert_slices_equivalent(SLC['b':'d':-1], SLC[:0])
-
-        assert_slices_equivalent(SLC[('c', 2)::-1], SLC[10::-1])
-        assert_slices_equivalent(SLC[:('c', 2):-1], SLC[:9:-1])
-        assert_slices_equivalent(SLC[('e', 0):('c', 2):-1], SLC[16:9:-1])
-
-    def test_multiindex_slice_first_level(self):
-        # GH 12697
-        freq = ['a', 'b', 'c', 'd']
-        idx = MultiIndex.from_product([freq, np.arange(500)])
-        df = DataFrame(list(range(2000)), index=idx, columns=['Test'])
-        df_slice = df.loc[pd.IndexSlice[:, 30:70], :]
-        result = df_slice.loc['a']
-        expected = DataFrame(list(range(30, 71)),
-                             columns=['Test'], index=range(30, 71))
-        tm.assert_frame_equal(result, expected)
-        result = df_slice.loc['d']
-        expected = DataFrame(list(range(1530, 1571)),
-                             columns=['Test'], index=range(30, 71))
-        tm.assert_frame_equal(result, expected)
-
     def test_multiindex_symmetric_difference(self):
         # GH 13490
         idx = MultiIndex.from_product([['a', 'b'], ['A', 'B']],
@@ -800,23 +756,6 @@ class TestMultiIndexBasic(object):
         # in theory should be inserting in a sorted space????
         frame.loc[('bar', 'three'), 'B'] = 0
         assert frame.sort_index().loc[('bar', 'three'), 'B'] == 0
-
-    def test_int_series_slicing(
-            self, multiindex_year_month_day_dataframe_random_data):
-        ymd = multiindex_year_month_day_dataframe_random_data
-        s = ymd['A']
-        result = s[5:]
-        expected = s.reindex(s.index[5:])
-        tm.assert_series_equal(result, expected)
-
-        exp = ymd['A'].copy()
-        s[5:] = 0
-        exp.values[5:] = 0
-        tm.assert_numpy_array_equal(s.values, exp.values)
-
-        result = ymd[5:]
-        expected = ymd.reindex(s.index[5:])
-        tm.assert_frame_equal(result, expected)
 
     @pytest.mark.parametrize('unicode_strings', [True, False])
     def test_mixed_depth_get(self, unicode_strings):
